menuentryswapper plugin: Fix Shift-click color bug
Prior to this patch, the "Shift-click" target used in the `MenuEntry`
added to the right-click menu of the inventory tab would be saved as
"<col=ff9040><col=ff9040>Shift-click</col>" due to being prepended with
the tag in MenuEntrySwapperPlugin and being wrapped with the tag in
WidgetMenuOption. This patch centralizes that logic in the added public
static function `WidgetMenuOption#applyTargetMenuColor`.