api: modify ChatMessage event to by backed by MessageNode
Keeping copies of the MessageNode's fields caused bugs due to the
MessageNode's message being updated but not the event's.

This fixes the emoji plugin causing the chat recolors to not work due to
applying the emojis to the event's message and not the message node's,
which had been previously recolored by the chat message manager.