xptracker: better support offline xp gains
This offsets the start xp by the gained xp when offline to keep the xp
per hour calculation correct.

This also changes initialization of the overall tracker to a game tick
after login, instead of after construction is set, which fixes the
xptracker not reinitializing properly if the plugin is flipped on when
already logged in.