banktags: don't use bank title for tab activation
This was parsing the currently searched tag out of the bank title when
the input text variable was something other than the search input. This
was breaking due to the bank plugin now changing the bank title of
searches.

Intead use the remembered search when this happens, since we have to
keep it around anyway to re-search for the banktag *after*
meslayer_close is called, in order to re-set the input text variable.